refactor(/controller): 예외 메시지의 내부 클래스명을 더 명확히 해라
