+++ src/main/java/racingcar/controller/ExceptionMessages.java
@@ -1,7 +1,7 @@
 package racingcar.controller;
 
 public class ExceptionMessages {
-    public class Default {
+    public class DefaultInputValidator {
         public static final String EMPTY_CAR_NAME = "자동차의 이름은 빈 문자열이어서는 안됩니다.";
         public static final String CAR_NAME_ONLY_DELIMITER = "자동차의 이름은 구분자인 쉼표(,)로만 이루어질 수 없습니다.";
         public static final String CAR_NAME_ENDS_WITH_DELIMITER = "자동차의 이름은 구분자인 쉼표(,)로 끝나서는 안됩니다.";
