+++ src/test/java/racingcar/controller/DefaultInputValidatorGameInputValidatorTest.java
@@ -2,19 +2,19 @@ package racingcar.controller;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
-import static racingcar.controller.ExceptionMessages.Default.CAR_NAME_ENDS_WITH_DELIMITER;
-import static racingcar.controller.ExceptionMessages.Default.CAR_NAME_ONLY_DELIMITER;
-import static racingcar.controller.ExceptionMessages.Default.EMPTY_CAR_NAME;
-import static racingcar.controller.ExceptionMessages.Default.EMPTY_TOTAL_ROUNDS;
-import static racingcar.controller.ExceptionMessages.Default.INVALID_CAR_NAME_CHARACTERS;
-import static racingcar.controller.ExceptionMessages.Default.TOTAL_ROUNDS_NOT_INTEGER;
+import static racingcar.controller.ExceptionMessages.DefaultInputValidator.CAR_NAME_ENDS_WITH_DELIMITER;
+import static racingcar.controller.ExceptionMessages.DefaultInputValidator.CAR_NAME_ONLY_DELIMITER;
+import static racingcar.controller.ExceptionMessages.DefaultInputValidator.EMPTY_CAR_NAME;
+import static racingcar.controller.ExceptionMessages.DefaultInputValidator.EMPTY_TOTAL_ROUNDS;
+import static racingcar.controller.ExceptionMessages.DefaultInputValidator.INVALID_CAR_NAME_CHARACTERS;
+import static racingcar.controller.ExceptionMessages.DefaultInputValidator.TOTAL_ROUNDS_NOT_INTEGER;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class DefaultGameInputValidatorTest {
+public class DefaultInputValidatorGameInputValidatorTest {
 
     private DefaultGameInputValidator gameInputValidator;
